% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nombre.R
\name{longitud_nombre}
\alias{longitud_nombre}
\title{Cuantifica la cantidad de caracteres que se requiere ubicar en la ventana de texto del archivo EDUS}
\usage{
longitud_nombre(n = "A")
}
\arguments{
\item{n}{corresponde al string de texto que contiene el nombre a medir}
}
\value{
a frase que indica la cantidad de caracteres y la posible necesidad de fraccionar el texto en dos lineas
}
\description{
Cuantifica la cantidad de caracteres que se requiere ubicar en la ventana de texto del archivo EDUS
}
\examples{
longitud_nombre("Javier Garcia Prado")
          longitud_nombre("Emilia Zamora")



}
