% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Identificacion.R
\name{tipo_identificacion}
\alias{tipo_identificacion}
\title{Comprueba longitud del tipo de identificacion}
\usage{
tipo_identificacion(x = 0)
}
\arguments{
\item{x}{valor cuya longitud se procesara para identificar similitud de cedula o de interno}
}
\value{
a frase que caracteriza si es una cedula, un interno o no se ajusta a ninguno
}
\description{
Comprueba longitud del tipo de identificacion
}
\examples{
tipo_identificacion(108360777)
          tipo_identificacion(1234100000678)
}
